fix: fetch the campaigns in the right order and with the right query
